:art: move icons to separate folder
