+++ src/components/icons/icon-add.jsx
