+++ src/components/icons/icon-close.jsx
