+++ src/components/icons/icon-purchase.jsx
