+++ src/components/my-sticker-packs.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 import StickerPack from './sticker-pack';
 import Sticker from './sticker';
 import parseResponse from '../parse-response';
-import IconAdd from './icon-add';
+import IconAdd from './icons/icon-add';
 
 class MyStickerPacks extends Component {
   static propTypes = {
