+++ src/components/sticker-pack-preview.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import parseResponse from '../parse-response';
-import IconPurchase from './icon-purchase';
-import IconClose from './icon-close';
+import IconPurchase from './icons/icon-purchase';
+import IconClose from './icons/icon-close';
 
 class StickerPackPreview extends Component {
   static propTypes = {
